requirements.txt
Flask==3.0.0
flask-wtf==1.2.1
flask-cors==4.0.0
pillow==10.1.0
python-telegram-bot==20.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
